organize src dir
