+++ src/Index.js
@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDom from 'react-dom'
 import { Router } from '@reach/router'
 
-import Main from './Main'
-import JobsList from './JobsList'
+import Main from './container/Main'
+import JobsList from './jobs/List'
 
 import 'bootstrap/dist/css/bootstrap.min.css'
