+++ src/index.html
@@ -8,7 +8,7 @@
 
   <body>
     <div id='root'></div>
-    <script src="./App.js"></script>
+    <script src="./Index.js"></script>
   </body>
 
 </html>
